// api-v2/src/grpc/proto/category/Option.ts
// Original file: node_modules/soum-proto/proto/category.proto

export interface Option {
  id?: string;
  nameEn?: string;
  nameAr?: string;
  status?: string;
  positionAr?: number;
  positionEn?: number;
}

export interface Option__Output {
  id: string;
  nameEn: string;
  nameAr: string;
  status: string;
  positionAr: number;
  positionEn: number;
}